# agricultural/api-tests/apifox-openapi-import.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "助农电商平台 API",
    "description": "用户端 / 商家端 / 管理员端 RESTful 接口 - 助农电商平台",
    "contact": {
      "name": "助农平台技术支持",
      "email": "support@agricultural.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/user/merchants/{id}/products": {
      "get": {
        "tags": ["用户端-商家"],
        "summary": "商家商品列表",
        "operationId": "merchantProducts",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultProductSimpleVO"}}}}
        }
      }
    },
    "/user/cart": {
      "get": {
        "tags": ["用户端-购物车"],
        "summary": "查看购物车",
        "operationId": "cartList",
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/CartItemVO"}}}}}
        }
      },
      "post": {
        "tags": ["用户端-购物车"],
        "summary": "添加商品到购物车",
        "operationId": "addToCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {"type": "integer"},
                  "productNum": {"type": "integer"}
                },
                "required": ["productId", "productNum"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      },
      "delete": {
        "tags": ["用户端-购物车"],
        "summary": "清空购物车",
        "operationId": "clearCart",
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/user/cart/{id}": {
      "put": {
        "tags": ["用户端-购物车"],
        "summary": "修改购物车商品数量",
        "operationId": "updateCart",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productNum": {"type": "integer"}
                },
                "required": ["productNum"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      },
      "delete": {
        "tags": ["用户端-购物车"],
        "summary": "删除购物车商品",
        "operationId": "deleteCart",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/user/addresses": {
      "get": {
        "tags": ["用户端-收货地址"],
        "summary": "地址列表",
        "operationId": "addressList",
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/UserAddressVO"}}}}}
        }
      },
      "post": {
        "tags": ["用户端-收货地址"],
        "summary": "添加地址",
        "operationId": "addAddress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiver": {"type": "string"},
                  "phone": {"type": "string"},
                  "province": {"type": "string"},
                  "city": {"type": "string"},
                  "county": {"type": "string"},
                  "town": {"type": "string"},
                  "detailAddress": {"type": "string"},
                  "isDefault": {"type": "integer"}
                },
                "required": ["receiver", "phone", "province", "city", "county", "detailAddress"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/user/addresses/{id}": {
      "put": {
        "tags": ["用户端-收货地址"],
        "summary": "修改地址",
        "operationId": "updateAddress",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "receiver": {"type": "string"},
                  "phone": {"type": "string"},
                  "province": {"type": "string"},
                  "city": {"type": "string"},
                  "county": {"type": "string"},
                  "town": {"type": "string"},
                  "detailAddress": {"type": "string"},
                  "isDefault": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      },
      "delete": {
        "tags": ["用户端-收货地址"],
        "summary": "删除地址",
        "operationId": "deleteAddress",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/user/addresses/{id}/default": {
      "put": {
        "tags": ["用户端-收货地址"],
        "summary": "设为默认地址",
        "operationId": "setDefaultAddress",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/user/after-sale": {
      "get": {
        "tags": ["用户端-售后"],
        "summary": "售后列表",
        "operationId": "afterSaleList",
        "parameters": [
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultAfterSaleVO"}}}}
        }
      },
      "post": {
        "tags": ["用户端-售后"],
        "summary": "发起售后",
        "operationId": "applyAfterSale",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "orderNo": {"type": "string"},
                  "afterSaleType": {"type": "integer", "description": "1-退款 2-退货 3-换货"},
                  "applyReason": {"type": "string"},
                  "proofImgUrls": {"type": "string"}
                },
                "required": ["orderNo", "afterSaleType", "applyReason"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/user/after-sale/{id}/escalate": {
      "put": {
        "tags": ["用户端-售后"],
        "summary": "申请管理员介入",
        "operationId": "escalateAfterSale",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/user/after-sale/{no}/messages": {
      "get": {
        "tags": ["用户端-售后沟通"],
        "summary": "沟通列表",
        "operationId": "afterSaleMessages",
        "parameters": [
          {"name": "no", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/AfterSaleCommunication"}}}}}
        }
      },
      "post": {
        "tags": ["用户端-售后沟通"],
        "summary": "发送消息",
        "operationId": "sendAfterSaleMessage",
        "parameters": [
          {"name": "no", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {"type": "string"}
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/merchant/products": {
      "get": {
        "tags": ["商家端-商品管理"],
        "summary": "商品列表",
        "operationId": "merchantProducts",
        "parameters": [
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultProductInfo"}}}}
        }
      },
      "post": {
        "tags": ["商家端-商品管理"],
        "summary": "新增商品",
        "operationId": "addMerchantProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductSaveDTO"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultProductInfo"}}}}
        }
      }
    },
    "/merchant/products/{id}": {
      "put": {
        "tags": ["商家端-商品管理"],
        "summary": "编辑商品",
        "operationId": "updateMerchantProduct",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductSaveDTO"
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      },
      "delete": {
        "tags": ["商家端-商品管理"],
        "summary": "删除商品",
        "operationId": "deleteMerchantProduct",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/merchant/products/{id}/qrcode": {
      "post": {
        "tags": ["商家端-商品管理"],
        "summary": "生成溯源码",
        "operationId": "generateProductQrcode",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultMapStringString"}}}}
        }
      }
    },
    "/merchant/products/{id}/status": {
      "put": {
        "tags": ["商家端-商品管理"],
        "summary": "上下架",
        "operationId": "updateProductStatus",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {"type": "integer", "description": "1-上架 2-下架"}
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/admin/merchants/{id}/status": {
      "put": {
        "tags": ["管理员端-商家管理"],
        "summary": "更新商家状态",
        "operationId": "updateMerchantStatus",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {"type": "integer", "description": "0-未开业 1-正常营业 2-禁用"}
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"type": "object"}
        }
      },
...
      "ProductSimpleVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "productName": {"type": "string"},
          "price": {"type": "number"},
          "productImg": {"type": "string"},
          "salesVolume": {"type": "integer"},
          "stock": {"type": "integer"}
        }
      },
      "PageResultProductSimpleVO": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/ProductSimpleVO"}}
        }
      },
...
      "CartItemVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "productId": {"type": "integer", "format": "int64"},
          "productName": {"type": "string"},
          "productImg": {"type": "string"},
          "productPrice": {"type": "number"},
          "productNum": {"type": "integer"},
          "stock": {"type": "integer"}
        }
      },
...
      "UserAddressVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "receiver": {"type": "string"},
          "phone": {"type": "string"},
          "province": {"type": "string"},
          "city": {"type": "string"},
          "county": {"type": "string"},
          "town": {"type": "string"},
          "detailAddress": {"type": "string"},
          "isDefault": {"type": "integer"}
        }
      },
      "AfterSaleVO": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "afterSaleNo": {"type": "string"},
          "orderNo": {"type": "string"},
          "afterSaleType": {"type": "integer"},
          "applyReason": {"type": "string"},
          "afterSaleStatus": {"type": "integer"},
          "handleResult": {"type": "string"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "PageResultAfterSaleVO": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/AfterSaleVO"}}
        }
      },
      "AfterSaleCommunication": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "afterSaleNo": {"type": "string"},
          "senderType": {"type": "integer", "description": "1-用户 2-商家 3-管理员"},
          "senderId": {"type": "integer", "format": "int64"},
          "content": {"type": "string"},
          "isRead": {"type": "integer"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
...
      "ProductSaveDTO": {
        "type": "object",
        "properties": {
          "productName": {"type": "string"},
          "categoryId": {"type": "integer", "format": "int64"},
          "price": {"type": "number"},
          "stock": {"type": "integer"},
          "stockWarning": {"type": "integer"},
          "productImg": {"type": "string"},
          "productDesc": {"type": "string"},
          "originPlace": {"type": "string"},
          "plantingCycle": {"type": "string"},
          "originPlaceDetail": {"type": "string"},
          "fertilizerType": {"type": "string"},
          "storageMethod": {"type": "string"},
          "transportMethod": {"type": "string"}
        }
      },
      "ProductInfo": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "productName": {"type": "string"},
          "merchantId": {"type": "integer", "format": "int64"},
          "categoryId": {"type": "integer", "format": "int64"},
          "price": {"type": "number"},
          "stock": {"type": "integer"},
          "stockWarning": {"type": "integer"},
          "productImg": {"type": "string"},
          "productDesc": {"type": "string"},
          "originPlace": {"type": "string"},
          "status": {"type": "integer", "description": "0-待审核 1-已上架 2-已下架 3-已驳回"},
          "salesVolume": {"type": "integer"},
          "score": {"type": "number"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "ResultProductInfo": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/ProductInfo"}
        }
      },
      "PageResultProductInfo": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/ProductInfo"}}
        }
      },
      "ResultMapStringString": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"type": "object", "additionalProperties": {"type": "string"}}
        }
      },
...
    }
  }
}